+++ examples/piggybank/src/Version1.tsx
@@ -1,6 +1,7 @@
 /* eslint-disable no-console */
 import React, { useEffect, useState, useContext, useRef } from 'react';
-import { JsonRpcClient, toBuffer } from '@concordium/web-sdk';
+import { toBuffer } from '@concordium/web-sdk';
+import { detectConcordiumProvider } from '@concordium/browser-wallet-api-helpers';
 import { smash, deposit, state, CONTRACT_NAME } from './utils';
 
 import PiggyIcon from './assets/piggy-bank-solid.svg';
@@ -16,59 +17,56 @@ const CONTRACT_INDEX = 81n; // V1 instance
 /** Should match the subindex of the instance targeted. */
 const CONTRACT_SUB_INDEX = 0n;
 
-interface Props {
-    client: JsonRpcClient;
+async function updateState(setSmashed: (x: boolean) => void, setAmount: (x: bigint) => void): Promise<void> {
+    const provider = await detectConcordiumProvider();
+    const res = await provider.getJsonRpcClient().invokeContract({
+        method: `${CONTRACT_NAME}.view`,
+        contract: { index: CONTRACT_INDEX, subindex: CONTRACT_SUB_INDEX },
+    });
+    if (!res || res.tag === 'failure' || !res.returnValue) {
+        throw new Error(`Expected succesful invocation`);
+    }
+    setSmashed(!!Number(res.returnValue.substring(0, 2)));
+    setAmount(toBuffer(res.returnValue.substring(2), 'hex').readBigUInt64LE(0) as bigint);
 }
 
-function updateState(client: JsonRpcClient, setSmashed: (x: boolean) => void, setAmount: (x: bigint) => void) {
-    client
-        .invokeContract({
-            method: `${CONTRACT_NAME}.view`,
-            contract: { index: CONTRACT_INDEX, subindex: CONTRACT_SUB_INDEX },
-        })
-        .then((res) => {
-            if (!res || res.tag === 'failure' || !res.returnValue) {
-                throw new Error(`Expected succesful invocation`);
-            }
-            setSmashed(!!Number(res.returnValue.substring(0, 2)));
-            setAmount(toBuffer(res.returnValue.substring(2), 'hex').readBigUInt64LE(0) as bigint);
-        });
-}
-
-export default function PiggyBank({ client }: Props) {
-    const { account, isConnected, jsonRpcUrl } = useContext(state);
+export default function PiggyBank() {
+    const { account, isConnected } = useContext(state);
     const [owner, setOwner] = useState<string>();
     const [smashed, setSmashed] = useState<boolean>();
     const [amount, setAmount] = useState<bigint>(0n);
     const input = useRef<HTMLInputElement>(null);
 
     useEffect(() => {
-        // Get piggy bank owner.
-        client.getInstanceInfo({ index: CONTRACT_INDEX, subindex: CONTRACT_SUB_INDEX }).then((info) => {
-            if (info?.name !== `init_${CONTRACT_NAME}`) {
-                // Check that we have the expected instance.
-                throw new Error(`Expected instance of PiggyBank: ${info?.name}`);
-            }
+        if (isConnected) {
+            // Get piggy bank owner.
+            detectConcordiumProvider()
+                .then((provider) =>
+                    provider.getJsonRpcClient().getInstanceInfo({ index: CONTRACT_INDEX, subindex: CONTRACT_SUB_INDEX })
+                )
+                .then((info) => {
+                    if (info?.name !== `init_${CONTRACT_NAME}`) {
+                        // Check that we have the expected instance.
+                        throw new Error(`Expected instance of PiggyBank: ${info?.name}`);
+                    }
 
-            setOwner(info.owner.address);
-        });
-    }, []);
+                    setOwner(info.owner.address);
+                });
+        }
+    }, [isConnected]);
 
     // The internal state of the piggy bank, which is either intact or smashed.
     useEffect(() => {
-        updateState(client, setSmashed, setAmount);
-    }, []);
+        if (isConnected) {
+            updateState(setSmashed, setAmount);
+        }
+    }, [isConnected]);
 
     // Disable use if we're not connected or if piggy bank has already been smashed.
     const canUse = isConnected && smashed !== undefined && !smashed;
 
     return (
-        <main className="piggybank">
-            <div className={`connection-banner ${isConnected ? 'connected' : ''}`}>
-                {isConnected ? `Connected: ${account}` : 'No wallet connection'}
-            </div>
-            <div>{jsonRpcUrl ? `JSON-RPC Url: ${jsonRpcUrl}` : 'No JSON-RPC Url yet'}</div>
-            <br />
+        <>
             {owner === undefined ? (
                 <div>Loading piggy bank...</div>
             ) : (
@@ -81,7 +79,7 @@ export default function PiggyBank({ client }: Props) {
                     </div>
                     <br />
                     <div>State: {smashed ? 'Smashed' : 'Intact'}</div>
-                    <button type="button" onClick={() => updateState(client, setSmashed, setAmount)}>
+                    <button type="button" onClick={() => updateState(setSmashed, setAmount)}>
                         ↻
                     </button>
                 </>
@@ -113,6 +111,6 @@ export default function PiggyBank({ client }: Props) {
             >
                 <HammerIcon width="40" />
             </button>
-        </main>
+        </>
     );
 }