+++ packages/browser-wallet-api-helpers/src/wallet-api-types.ts
@@ -2,6 +2,7 @@ import type {
     AccountTransactionPayload,
     AccountTransactionSignature,
     AccountTransactionType,
+    JsonRpcClient,
     SchemaVersion,
 } from '@concordium/web-sdk';
 
@@ -75,7 +76,14 @@ interface MainWalletApi {
      */
     connect(): Promise<string | undefined>;
 
+    /**
+     * Returns some connected account, prioritizing the most recently selected. Resolves with account address or undefined if there are no connected account.
+     */
+    getMostRecentlySelectedAccount(): Promise<string | undefined>;
+
     removeAllListeners(event?: EventType | string | undefined): this;
+
+    getJsonRpcClient(): JsonRpcClient;
 }
 
 export type WalletApi = MainWalletApi & EventListeners;