+++ packages/browser-wallet-api-helpers/CHANGELOG.md
@@ -2,6 +2,13 @@
 
 ## 0.2.0
 
+### Added
+
+-   Expose a JSON-RPC client, using the wallet's current JSON-RPC server.
+-   `getMostRecentlySelectedAccount` method. This method allows dApps to get the most prioritized account without using `connect`. In a future release it will be updated to actually return the most recently selected account.
+
+### (Breaking) Changed
+
 -   Updated API of sendTransaction and signMessage to require the account address.
 -   Updated API to include an 'accountDisconnected' event.